feat: added customizable refit preset support
How to use:

It is now possible to create a global override of the baked in refit
preset. To do that simply have a preset named either "Female-Refit" for
female actors or "Male-Refit" for male actors. If not having such a
preset in the preset list OBody will fall back to the baked in one.

It is also possible to create refit presets for each normal presets.
Say you have a preset called "MyPreset". Now if you have another preset
called "MyPreset-Refit" each character that has "MyPreset" as selected
body will use "MyPreset-Refit" while clothed.

It is very important that refit presets must only contain the
differences from the normal base preset. Say that "MyPreset" has
value 10 for the Pushup slider. If your desired value for the Pushup
slider is 50 then "MyPreset-Refit" must contain value 40 (50-10) for
the Pushup slider.

Refit presets (having a suffix "-Refit" in their names) are blacklisted
by default meaning they won't be included in random body distribution
or appear in the OBody preset selection menu. They are solely used for
the refit feature.